add resources

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/service/contract/SessionPayload.java
@@ -10,7 +10,7 @@ public class SessionPayload implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private long sessionId;
+    private Integer sessionId;
 
     private LocalDateTime startedDateTime;
